+++ src/main/java/wooteco/subway/repository/SectionRepository.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dao;
+package wooteco.subway.repository;
 
 import java.util.List;
 import javax.sql.DataSource;
