+++ src/main/java/wooteco/subway/repository/StationRepository.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dao;
+package wooteco.subway.repository;
 
 import java.util.List;
 import java.util.Optional;
